Safari-609.1.11.1/LayoutTests/fast/forms/ios/scroll-to-reveal-focused-select-expected.txt


Verifies that focusing a select element after user interaction scrolls to reveal the select element. To manually test, focus the top select element and tap the red box to focus the bottom select element. The bottom select element should be scrolled into view.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.activeElement is document.getElementById('bottom')
PASS pageYOffset is >= 1000
PASS successfullyParsed is true

TEST COMPLETE
